pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "energize_denver_eaas"
version = "0.1.0"
description = "Energize Denver Risk & Retrofit Strategy Platform"
authors = [{name = "Robert Padgett", email = "rpadgett@clms.com"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-storage>=2.10.0",
    "google-cloud-bigquery>=3.11.0",
    "pandas>=2.0.0",
    "numpy>=1.26.0",
    "pyarrow>=14.0.0",
    "db-dtypes>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
viz = [
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "folium>=0.14.0",
    "plotly>=5.14.0",
]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
